requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Werkzeug
bcrypt
openai
python-dotenv
# deepinfra
Flask-Migrate
gunicorn
psycopg2-binary
boto3
Flask-Restful